Properly support server errors in PG timeout query executor

+++ postbound/db/postgres.py
@@ -73,7 +73,7 @@ from ..qal import (
     WindowExpression,
     transform,
 )
-from ..util import StateError, Version
+from ..util import StateError, Version, jsondict
 from ._db import (
     Database,
     DatabasePool,
@@ -721,41 +721,42 @@ class PostgresInterface(Database):
 
         if cache_enabled and query in self._query_cache:
             query_result = self._query_cache[query]
-        else:
-            try:
-                self._current_geqo_state = self._obtain_geqo_state()
-                prepared_query, deactivated_geqo = self._prepare_query_execution(query)
-                self._cursor.execute(prepared_query)
-                query_result = (
-                    self._cursor.fetchall() if self._cursor.rowcount >= 0 else None
-                )
-                if deactivated_geqo:
-                    self._restore_geqo_state()
-                if cache_enabled:
-                    self._inflate_query_cache()
-                    self._query_cache[prepared_query] = query_result
-            except (psycopg.InternalError, psycopg.OperationalError) as e:
-                msg = "\n".join(
-                    [
-                        f"At {util.timestamp()}",
-                        "For query:",
-                        str(query),
-                        "Message:",
-                        str(e),
-                    ]
-                )
-                raise DatabaseServerError(msg, e)
-            except psycopg.Error as e:
-                msg = "\n".join(
-                    [
-                        f"At {util.timestamp()}",
-                        "For query:",
-                        str(query),
-                        "Message:",
-                        str(e),
-                    ]
-                )
-                raise DatabaseUserError(msg, e)
+            return query_result if raw else _simplify_result_set(query_result)
+
+        try:
+            self._current_geqo_state = self._obtain_geqo_state()
+            prepared_query, deactivated_geqo = self._prepare_query_execution(query)
+            self._cursor.execute(prepared_query)
+            query_result = (
+                self._cursor.fetchall() if self._cursor.rowcount >= 0 else None
+            )
+            if deactivated_geqo:
+                self._restore_geqo_state()
+            if cache_enabled:
+                self._inflate_query_cache()
+                self._query_cache[prepared_query] = query_result
+        except (psycopg.InternalError, psycopg.OperationalError) as e:
+            msg = "\n".join(
+                [
+                    f"At {util.timestamp()}",
+                    "For query:",
+                    str(query),
+                    "Message:",
+                    str(e),
+                ]
+            )
+            raise DatabaseServerError(msg, e)
+        except psycopg.Error as e:
+            msg = "\n".join(
+                [
+                    f"At {util.timestamp()}",
+                    "For query:",
+                    str(query),
+                    "Message:",
+                    str(e),
+                ]
+            )
+            raise DatabaseUserError(msg, e)
 
         return query_result if raw else _simplify_result_set(query_result)
 
@@ -801,7 +802,7 @@ class PostgresInterface(Database):
         """
         return self._connection.info.backend_pid
 
-    def describe(self) -> dict:
+    def describe(self) -> jsondict:
         base_info = {
             "system_name": self.database_system_name(),
             "system_version": self.database_system_version(),
@@ -821,22 +822,29 @@ class PostgresInterface(Database):
             if setting in _SignificantPostgresSettings
         }
 
-        schema_info: list = []
-        for table in self.schema().tables():
-            column_info: list = []
-            for column in self.schema().columns(table):
+        schema_info: list[jsondict] = []
+        for table in self._db_schema.tables():
+            column_info: list[jsondict] = []
+
+            for column in self._db_schema.columns(table):
                 column_info.append(
-                    {"column": str(column), "indexed": self.schema().has_index(column)}
+                    {
+                        "column": str(column),
+                        "indexed": self.schema().has_index(column),
+                        "foreign_keys": self._db_schema.foreign_keys_on(column),
+                    }
                 )
+
             schema_info.append(
                 {
                     "table": str(table),
                     "n_rows": self.statistics().total_rows(table, emulated=True),
                     "columns": column_info,
+                    "primary_key": self._db_schema.primary_key_column(table).name,
                 }
             )
-        base_info["schema_info"] = schema_info
 
+        base_info["schema_info"] = schema_info
         return base_info
 
     def reset_connection(self) -> int:
@@ -3153,12 +3161,14 @@ class ParallelQueryExecutor:
 def _timeout_query_worker(
     query: qal.SqlQuery | str,
     pg_instance: PostgresInterface,
-    query_result_sender: mp_conn.Connection,
+    result_send: mp_conn.Connection,
+    err_send: mp_conn.Connection,
     **kwargs,
 ) -> None:
     """Internal function to the `TimeoutQueryExecutor` to run individual queries.
 
-    Query results are sent via the pipe, not as a return value.
+    Query results are sent via the `result_send` pipe, not as a return value. In case of any errors, these are sent via the
+    `err_send` pipe. Therefore, it is best to check the `err_send` pipe first, before reading from the `result_send` pipe.
 
     Parameters
     ----------
@@ -3166,13 +3176,18 @@ def _timeout_query_worker(
         Query to execute
     pg_instance : PostgresInterface
         Database connection to execute the query on
-    query_result_sender : mp_conn.Connection
+    result_send : mp_conn.Connection
         Pipe connection to send the query result
+    err_send : mp_conn.Connection
+        Pipe connection to send any errors that occurred during the query execution
     kwargs : Any
         Additional parameters to pass to the `PostgresInterface.execute_query` method.
     """
-    result = pg_instance.execute_query(query, **kwargs)
-    query_result_sender.send(result)
+    try:
+        result = pg_instance.execute_query(query, **kwargs)
+        result_send.send(result)
+    except Exception as e:
+        err_send.send(e)
 
 
 class TimeoutQueryExecutor:
@@ -3231,36 +3246,52 @@ class TimeoutQueryExecutor:
         PostgresInterface.execute_query
         PostgresInterface.reset_connection
         """
-        query_result_receiver, query_result_sender = mp.Pipe(False)
+        result_recv, result_send = mp.Pipe(False)
+        error_recv, error_send = mp.Pipe(False)
         query_execution_worker = mp.Process(
             target=_timeout_query_worker,
-            args=(query, self._pg_instance, query_result_sender),
+            args=(query, self._pg_instance, result_send, error_send),
             kwargs=kwargs,
         )
 
         query_execution_worker.start()
         query_execution_worker.join(timeout)
+
+        # We perform the timeout check before doing anything else to make sure that the worker process cannot terminate
+        # immediately after the timeout has been reached. E.g., suppose that  the query is still running after calling join().
+        # If we would now proceed to check the error-pipe, the query would have more time to terminate while we are performing
+        # our error checks. This might result in an involuntary increase in the timeout duration. By keeping the timeout check
+        # as close to the join() call as possible, we minimize this risk.
         timed_out = query_execution_worker.is_alive()
 
+        # Now that we know whether the worker timed out or not, we need to make sure that it actually terminated properly
+        # (or timed out). In case of an error, we just propagate it to the client.
+        if error_recv.poll():
+            raise error_recv.recv()
+
+        # At this point we know that the worker either terminated in time or that it timed out, but it did not error.
+        # Both the timeout and the termination case can be handled in a pretty straightforward manner. The only important thing
+        # is that in case of a timeout, we reset the connection to ensure that there is no dangling or inconsistent state left
+        # over.
         if timed_out:
             query_execution_worker.terminate()
             query_execution_worker.join()
             self._pg_instance.reset_connection()
             query_result = None
         else:
-            query_result = query_result_receiver.recv()
+            query_result = result_recv.recv()
 
         query_execution_worker.close()
-        query_result_sender.close()
-        query_result_receiver.close()
+        result_send.close()
+        result_recv.close()
 
         if timed_out:
             raise TimeoutError(query)
         else:
             return query_result
 
-    def __call__(self, query: qal.SqlQuery | str, timeout: float) -> Any:
-        return self.execute_query(query, timeout)
+    def __call__(self, query: qal.SqlQuery | str, timeout: float, **kwargs) -> Any:
+        return self.execute_query(query, timeout, **kwargs)
 
 
 PostgresExplainJoinNodes = {
